+++ cluster-autoscaler/core/scaledown/actuation/budgets.go
@@ -1,0 +1,49 @@
+/*
+Copyright 2022 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package actuation
+
+import (
+	apiv1 "k8s.io/api/core/v1"
+)
+
+// cropNodesToBudgets crops the provided node lists to respect scale-down max parallelism budgets.
+func (a *Actuator) cropNodesToBudgets(empty, needDrain []*apiv1.Node) ([]*apiv1.Node, []*apiv1.Node) {
+	emptyInProgress, drainInProgress := a.nodeDeletionTracker.DeletionsInProgress()
+	parallelismBudget := a.ctx.MaxScaleDownParallelism - len(emptyInProgress) - len(drainInProgress)
+	drainBudget := a.ctx.MaxDrainParallelism - len(drainInProgress)
+
+	var emptyToDelete []*apiv1.Node
+	for _, node := range empty {
+		if len(emptyToDelete) >= parallelismBudget {
+			break
+		}
+		emptyToDelete = append(emptyToDelete, node)
+	}
+
+	parallelismBudgetLeft := parallelismBudget - len(emptyToDelete)
+	drainBudget = min(parallelismBudgetLeft, drainBudget)
+
+	var drainToDelete []*apiv1.Node
+	for _, node := range needDrain {
+		if len(drainToDelete) >= drainBudget {
+			break
+		}
+		drainToDelete = append(drainToDelete, node)
+	}
+
+	return emptyToDelete, drainToDelete
+}
